+++ samples/kaspresso-sample/src/androidTest/java/com/kaspersky/kaspressample/docloc_tests/ChangeSysLanguageTestCase.kt
@@ -31,19 +31,15 @@ class ChangeSysLanguageTestCase : DocLocScreenshotTestCase(
     val activityTestRule = ActivityTestRule(MainActivity::class.java, true, false)
 
     @Test
-    fun showAndroidSettings() {
-        before {
-        }.after {
-        }.run {
-            step("Start Android OS Settings") {
-                val intent = Intent(Settings.ACTION_SETTINGS)
-                intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-                device.targetContext.startActivity(intent)
-                Thread.sleep(3000)
-            }
-            step("Capture Android OS Settings") {
-                captureScreenshot("android_setting")
-            }
+    fun showAndroidSettings() = run {
+        step("Start Android OS Settings") {
+            val intent = Intent(Settings.ACTION_SETTINGS)
+            intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
+            device.targetContext.startActivity(intent)
+            Thread.sleep(3000)
+        }
+        step("Capture Android OS Settings") {
+            captureScreenshot("android_setting")
         }
     }
 }
