+++ samples/kaspresso-sample/src/androidTest/java/com/kaspersky/kaspressample/docloc_tests/cutomdirectory/FlatDirectoryProvider.kt
@@ -1,0 +1,8 @@
+package com.kaspersky.kaspressample.docloc_tests.cutomdirectory
+
+import com.kaspersky.kaspresso.device.screenshots.screenshotfiles.ScreenshotDirectoryProvider
+import com.kaspersky.kaspresso.device.screenshots.screenshotfiles.TestMethod
+
+internal object FlatDirectoryProvider : ScreenshotDirectoryProvider {
+    override fun getDirectoryForTest(testMethod: TestMethod, runNumber: Int): String = ""
+}
