+++ wiki/07_DocLoc.md
@@ -210,3 +210,9 @@ There is also `getUiSafeProxyFromImplementation` which wraps an implementation r
 
 For full example, check [AdvancedScreenshotSampleTest](../samples/kaspresso-sample/src/androidTest/java/com/kaspersky/kaspressample/docloc_tests/advanced/AdvancedScreenshotSampleTest.kt) class.
  
+## Modifying screenshots path and name
+
+By default, all screenshots are stored at /sdcard/<screenshots root dir>/<locale>/<full qualified test class name>/<method name>. 
+You can change this behavior by providing custom [ScreenshotDirectoryProvider](../kaspresso/src/main/kotlin/com/kaspersky/kaspresso/device/screenshots/screenshotfiles/ScreenshotDirectoryProvider.kt) 
+and [ScreenshotNameProvider](../kaspresso/src/main/kotlin/com/kaspersky/kaspresso/device/screenshots/screenshotfiles/ScreenshotNameProvider.kt) implementations. 
+Find out details [here](../samples/kaspresso-sample/src/androidTest/java/com/kaspersky/kaspressample/docloc_tests/cutomdirectory/CustomDirectoryScreenshotSampleTest.kt)
